# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "gpt-terminal-plus API",
    "version": "0.1.0",
    "description": "OpenAPI generated from JSDoc annotations.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Public base URL"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_TOKEN"
      }
    },
    "schemas": {
      "ExecutionResult": {
        "type": "object",
        "properties": {
          "stdout": {
            "type": "string"
          },
          "stderr": {
            "type": "string"
          },
          "exitCode": {
            "type": "integer"
          },
          "error": {
            "type": "boolean"
          }
        },
        "required": [
          "stdout",
          "stderr",
          "exitCode",
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/command/execute-shell": {
      "post": {
        "operationId": "executeShell",
        "summary": "Execute a command (literal by default)",
        "description": "Default runs the binary with args via spawn (no shell expansion). If `shell` is provided (e.g., bash, powershell), the command runs under that shell.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "command": {
                    "type": "string"
                  },
                  "args": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "shell": {
                    "type": "string",
                    "description": "Optional shell to run under (bash, powershell, etc). If omitted, defaults to safe literal mode with no shell expansion."
                  }
                },
                "required": [
                  "command"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Execution complete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "$ref": "#/components/schemas/ExecutionResult"
                    },
                    "aiAnalysis": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "result"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "x-openai-isConsequential": false
      }
    },
...
    "/settings": {
      "get": {
        "operationId": "getSettings",
        "summary": "Get redacted configuration settings",
        "description": "Returns grouped configuration values with secrets redacted. Values overridden by environment variables are marked as readOnly.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Redacted settings grouped by category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "server": {
                      "type": "object",
                      "properties": {
                        "port": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "number"
                            },
                            "readOnly": {
                              "type": "boolean"
                            }
                          }
                        },
                        "httpsEnabled": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "boolean"
                            },
                            "readOnly": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    },
                    "security": {
                      "type": "object",
                      "properties": {
                        "apiToken": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "string"
                            },
                            "readOnly": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    },
                    "llm": {
                      "type": "object",
                      "properties": {
                        "provider": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "string"
                            },
                            "readOnly": {
                              "type": "boolean"
                            }
                          }
                        },
                        "openai.baseUrl": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "string"
                            },
                            "readOnly": {
                              "type": "boolean"
                            }
                          }
                        },
                        "openai.apiKey": {
                          "type": "object",
                          "properties": {
                            "value": {
                              "type": "string"
                            },
                            "readOnly": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "sample": {
                    "summary": "Example response",
                    "value": {
                      "server": {
                        "port": {
                          "value": 5005,
                          "readOnly": false
                        },
                        "httpsEnabled": {
                          "value": false,
                          "readOnly": false
                        }
                      },
                      "security": {
                        "apiToken": {
                          "value": "*****",
                          "readOnly": true
                        }
                      },
                      "llm": {
                        "provider": {
                          "value": "openai",
                          "readOnly": false
                        },
                        "openai.baseUrl": {
                          "value": "",
                          "readOnly": false
                        },
                        "openai.apiKey": {
                          "value": "*****",
                          "readOnly": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/config/schema": {
      "get": {
        "operationId": "getConfigSchema",
        "summary": "Get configuration schema",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Configuration schema"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/config/override": {
      "get": {
        "operationId": "getConfigOverride",
        "summary": "Get current configuration",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current configuration"
          }
        },
        "x-openai-isConsequential": false
      },
      "post": {
        "operationId": "setConfigOverride",
        "summary": "Update configuration",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Configuration updated"
          },
          "400": {
            "description": "Invalid configuration"
          }
        },
        "x-openai-isConsequential": false
      }
    },
    "/config/settings": {
      "get": {
        "operationId": "getConfigSettings",
        "summary": "Get redacted configuration settings",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Redacted configuration settings"
          }
        },
        "x-openai-isConsequential": false
      }
    },
...
  },
...
}